fix(security): SEC-REALTIME-EVENT-PII-BY-VALUE — rebase realtime seam onto canonical secret/Unicode detectors (#1618)
Delete the realtime-local duplicate secret/Unicode scanner
(friday-realtime-secret-unicode-scan.ts) and re-point the event-payload
redactor at the canonical shared detectors #1619 landed, called exactly as the
memory-egress + audit sinks do:
  - secret shapes -> findSecretShapeSpans / redactSecretShapesInString
                     (src/security/friday-secret-shape-redactor.ts)
  - Unicode fold  -> buildUnicodeDetectionCopy / redactUnicodeObfuscated
                     (src/security/friday-unicode-pii-normalizer.ts)

The secret path is a strict-superset substitution: the canonical detector adds
hf_/gsk_/glpat-/AWS/Google/Slack/npm shapes the duplicate list never had (RED
pre-rebase -> GREEN), with the byte-preserving credential-subspan mapping
unchanged (Bearer scheme / assignment label + quotes + fullwidth prefix survive
byte-for-byte). All #1618 round-7/8/9 tests stay GREEN — no regression.

The value-PII PRESERVING fold (round-9 F2b-ND-1: obfuscated PII coverage WITHOUT
ideographic-space / No-Nl over-redaction) has NO canonical equivalent — the
canonical aggressive NFKD fold folds U+3000->space and circled-digits->ASCII,
which would reintroduce exactly that benign over-redaction. To honor NO-DEGRADE
it is retained as a minimal, DETECTOR-FREE fold adapter
(friday-realtime-value-pii-fold.ts): zero secret patterns, PII classification
still delegated to the shared guard. Follow-up (out of scope; touches #1619):
add a fold-policy param to buildUnicodeDetectionCopy so this adapter collapses
into the shared module.
